feat: ensure only auth user data can be fetched

+++ src/factories/user.js
@@ -7,8 +7,14 @@ import {
   PostgresGetUserByIdRepository,
   PostgresGetUserBalanceRepository,
 } from '../repositories/postgres/index.js'
+
 // Adapter Factories
-import { PasswordHasherAdapter, UuidAdapter } from '../adapters/index.js'
+import {
+  PasswordHasherAdapter,
+  UuidAdapter,
+  TokenGeneratorAdapter,
+} from '../adapters/index.js'
+
 // Use-Cases Factories
 import {
   GetUserByIdUseCase,
@@ -17,8 +23,10 @@ import {
   UpdateUserUseCase,
   GetUserBalanceUseCase,
 } from '../use-cases/index.js'
+
 // Helpers Factories
 import { GetUserHelper } from '../helpers/http.js'
+
 // Controllers Factories
 import { UserController } from '../controllers/UserController.js'
 
@@ -36,25 +44,41 @@ export const makeGetUserController = () => {
   // Adapters
   const passwordHasherAdapter = new PasswordHasherAdapter()
   const uuidAdapter = new UuidAdapter()
+  const tokenGeneratorAdapter = new TokenGeneratorAdapter()
+  if (
+    !process.env.JWT_ACCESS_TOKEN_SECRET ||
+    !process.env.JWT_REFRESH_TOKEN_SECRET
+  ) {
+    throw new Error('JWT secrets are not configured')
+  }
+
   // Use-Cases
   const getUserByIdUseCase = new GetUserByIdUseCase(
     postgresGetUserByIdRepository,
   )
+
   const updateUserUseCase = new UpdateUserUseCase(
     postgresGetUserByEmailRepository,
     postgresUpdateUserRepository,
     passwordHasherAdapter,
   )
+
   const deleteUserUseCase = new DeleteUserUseCase(postgresDeleteUserRepository)
+
   const createUserUseCase = new CreateUserUseCase(
     postgresGetUserByEmailRepository,
     postgresCreateUserRepository,
     passwordHasherAdapter,
     uuidAdapter,
+    tokenGeneratorAdapter,
+    process.env.JWT_ACCESS_TOKEN_SECRET,
+    process.env.JWT_REFRESH_TOKEN_SECRET,
   )
+
   const getUserBalanceUseCase = new GetUserBalanceUseCase(
     postgresGetUserBalanceRepository,
   )
+
   // Helpers
   const getUserHelper = new GetUserHelper(getUserByIdUseCase)
